pyproject.toml
[tool.poetry]
name = "friend"
version = "0.1.0"
description = ""
authors = ["yoosunu <yoosunu010@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
torch = "^2.4.0"
transformers = "^4.43.2"
datasets = "^2.20.0"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
